(feat) can retrieve weather data and temperature by city

+++ Back/backend/migrations/20200310210133-create-precipitation.js
@@ -1,30 +1,30 @@
 'use strict';
 module.exports = {
-  up: (queryInterface, Sequelize) => {
-    return queryInterface.createTable('Precipitations', {
-      id: {
-        allowNull: false,
-        autoIncrement: true,
-        primaryKey: true,
-        type: Sequelize.INTEGER
-      },
-      mode: {
-        type: Sequelize.STRING
-      },
-      value: {
-        type: Sequelize.DECIMAL
-      },
-      createdAt: {
-        allowNull: false,
-        type: Sequelize.DATE
-      },
-      updatedAt: {
-        allowNull: false,
-        type: Sequelize.DATE
-      }
-    });
-  },
-  down: (queryInterface, Sequelize) => {
-    return queryInterface.dropTable('Precipitations');
-  }
+    up: (queryInterface, Sequelize) => {
+        return queryInterface.createTable('Precipitations', {
+            id: {
+                allowNull: false,
+                autoIncrement: true,
+                primaryKey: true,
+                type: Sequelize.INTEGER
+            },
+            mode: {
+                type: Sequelize.STRING
+            },
+            value: {
+                type: Sequelize.DECIMAL(10, 2)
+            },
+            createdAt: {
+                allowNull: false,
+                type: Sequelize.DATE
+            },
+            updatedAt: {
+                allowNull: false,
+                type: Sequelize.DATE
+            }
+        });
+    },
+    down: (queryInterface, Sequelize) => {
+        return queryInterface.dropTable('Precipitations');
+    }
 };
